+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/components/ratings/StarRatingView.kt
@@ -16,6 +16,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
@@ -30,6 +31,7 @@ import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.extensions.abbreviated
 import java.util.UUID
+import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @Composable
 fun StarRatingView(
@@ -62,9 +64,6 @@ private fun Content(
     state: RatingState.Loaded,
     onClick: () -> Unit
 ) {
-    if (state.noRatings) {
-        return
-    }
     Row(
         modifier = Modifier
             .padding(horizontal = 14.dp, vertical = 4.dp)
@@ -78,7 +77,10 @@ private fun Content(
         )
         state.total?.let {
             TextP40(
-                text = it.abbreviated,
+                text = when (it) {
+                    0 -> stringResource(LR.string.podcast_not_enough_ratings)
+                    else -> it.abbreviated
+                },
                 modifier = Modifier.padding(start = 6.dp)
             )
         }